+++ frontend/src/components/layout/MainLayout/index.tsx
@@ -1,16 +1,21 @@
 'use client';
 
-import CustomAppBar from '@/components/MainLayout/CustomAppBar';
-import {Box, Grid, Toolbar} from '@mui/material';
-import SideBar from '@/components/MainLayout/SideBar';
-import {io} from 'socket.io-client';
-import {useContext, useEffect, useState} from 'react';
-import {UserContext} from '../MainLayout/Context';
 import {useRouter} from 'next/router';
-import {UserType} from '@/types/UserContext';
-import apiManager from '@/api/apiManager';
+import {useContext, useEffect, useState} from 'react';
+
+import {io} from 'socket.io-client';
 import {useRecoilState, useSetRecoilState} from 'recoil';
+
+import {Box, Grid, Toolbar} from '@mui/material';
+
+import apiManager from '@/api/apiManager';
+import CustomAppBar from '@/components/layout/MainLayout/CustomAppBar';
+import SideBar from '@/components/layout/MainLayout/SideBar';
+import {UserContext} from '@/components/layout/MainLayout/Context';
+import {UserType} from '@/types/UserContext';
 import {dmList} from '@/states/dmUser';
+import {profileState} from '@/states/profile';
+import {socketState} from '@/states/sockets';
 
 interface MainLayoutProps {
   children: React.ReactNode;
@@ -24,6 +29,8 @@ function MainLayout({children}: MainLayoutProps) {
   const {block_users} = useContext(UserContext);
   const setDmList = useSetRecoilState(dmList);
   const router = useRouter();
+  const [profileDataState, setProfileDataState] = useRecoilState(profileState);
+  const [socketDataState, setSocketDataState] = useRecoilState(socketState);
 
   const [initMainLayout, setInitMainLayout] = useState(false);
 
@@ -38,17 +45,31 @@ function MainLayout({children}: MainLayoutProps) {
       try {
         const response = await apiManager.get('/user/user-info');
         console.log('response: ', response);
+        const {user_id, user_image, user_nickname, is_2fa_enabled} =
+          response.data;
+        setProfileDataState({
+          user_id,
+          image: user_image,
+          nickname: user_nickname,
+          is_2fa_enabled,
+        });
         setUserId(response.data.user_id);
         setUserNickName(response.data.user_nickname);
-        setUserImage(`${response.data.user_image}`);
-        const socketIo = io(`${process.env.NEXT_PUBLIC_BACKEND_SERVER}/chat`, {
-          query: {
-            user_id: response.data.user_id,
-            nickname: response.data.user_nickname,
-            user_image: `${response.data.user_image}`,
-          },
-        });
-        setChatSocket(socketIo);
+        setUserImage(response.data.user_image);
+        console.log(response.data);
+        const chatSocket = io(
+          `${process.env.NEXT_PUBLIC_BACKEND_SERVER}/chat`,
+          {
+            query: {
+              user_id: response.data.user_id,
+              nickname: response.data.user_nickname,
+              user_image: response.data.user_image,
+            },
+          }
+        );
+        console.log(chatSocket);
+        setChatSocket(chatSocket);
+        setSocketDataState({...socketDataState, chat: chatSocket});
         const rep_block_list = await apiManager.get(
           `/chatroom/block_list/${response.data.user_id}`
         );
